+++ crates/kawauso-project/src/lib.rs
@@ -1,1 +1,8 @@
 #![cfg_attr(not(doctest), doc = include_str!("../README.md"))]
+
+pub mod error;
+pub mod project;
+pub mod search;
+
+pub use self::project::Project;
+pub use self::search::Search;